+++ src/_mage.sh
@@ -240,31 +240,7 @@ case "${@}" in
   ;;
 
 "build"*)
-  # Initialize variables with default values
-  LANGS="en_US nl_NL"
-  JOBS=4
-  FORCE=""
-  DEFAULT_ARGS="$LANGS $JOBS"
-  ARGS=${@:2}
-
-  # If language is not found in ARGS add the default
-  if [[ $ARG != *_* ]]; then
-    ARGS="$LANGS $ARGS"
-  fi
-
-  # If Jobs is not found in ARGS add the default
-  if [[ "$ARGS" != *"-j"* ]] || [[ "$ARGS" != *"--jobs"* ]]; then
-    ARGS="$ARGS -j $JOBS"
-  fi
-
-  # If force is found in ARGS add it to both adminhtml and frontend
-  if [[ "$ARGS" == *"-f"* ]] || [[ "$ARGS" == *"--force"* ]]; then
-    FORCE="-f"
-  fi
-
-  # Deploy static content
-  $MAGENTO_CLI setup:static-content:deploy -a adminhtml en_US ${FORCE} &&
-  $MAGENTO_CLI setup:static-content:deploy -a frontend ${ARGS:-$DEFAULT_ARGS}
+  mage_build ${@:2}
   ;;
 
 "run"*)